requirements.txt
Django==1.9.6
MySQL-python==1.2.5
PyJWT==1.4.2
Pygments==2.1.3
argparse==1.2.1
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
dj-database-url==0.3.0
dj-static==0.0.6
django-cors-headers==2.0.0
django-social-auth==0.7.28
django-toolbelt==0.0.1
djangorestframework==3.2.4
enum34==1.1.6
google-api-python-client==1.6.0
gunicorn==19.3.0
httplib2==0.9.2
importlib==1.0.4
ipdb==0.10.1
ipython==5.1.0
ipython-genutils==0.1.0
oauth2==1.9.0.post1
oauth2client==4.0.0
oauthlib==2.0.1
pathlib2==2.1.0
pexpect==4.2.1
pickleshare==0.7.4
prompt-toolkit==1.0.9
psycopg2==2.6.1
ptyprocess==0.5.1
pyRFC3339==1.0
pyasn1==0.1.9
pyasn1-modules==0.0.8
python-openid==2.2.5
python-social-auth==0.3.3
pytz==2016.10
requests==2.12.4
requests-oauthlib==0.7.0
rsa==3.4.2
simplegeneric==0.8.1
six==1.10.0
social-auth-app-django==0.1.0
social-auth-core==0.2.1
static3==0.6.1
traitlets==4.3.1
uritemplate==3.0.0
wcwidth==0.1.7
wheel==0.24.0
whitenoise==2.0.2
wsgiref==0.1.2
